@@ -9,8 +9,8 @@
   },
   "author": "https://github.com/orgs/WebJamApps/teams/handlersandkickers",
   "engines": {
-    "node": ">=18.12.1 <19",
-    "npm": ">=8.19.2 <9"
+    "node": ">=18.13.0 <19",
+    "npm": ">=9.2.0 <10"
   },
   "files": [
     "webpack.config.js",
@@ -96,28 +91,26 @@
     "html-webpack-plugin": "^5.5.0",
     "jest-environment-jsdom": "^29.0.1",
     "jquery": "^3.6.1",
-    "jsonwebtoken": "^9.0.0",
+    "jwt-simple": "^0.5.6",
     "mini-css-extract-plugin": "^1.3.1",
     "moment": "^2.29.4",
     "popper.js": "^1.16.1",
     "process": "^0.11.10",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hot-loader": "^4.13.0",
-    "react-notifications-component": "^3.1.0",
-    "react-player": "^2.10.0",
-    "react-redux": "^8.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-notifications-component": "^4.0.1",
+    "react-player": "^2.11.0",
+    "react-redux": "^8.0.5",
     "react-resize-detector": "^7.1.2",
-    "react-router-dom": "^5.2.1",
-    "react-share": "^4.4.0",
+    "react-router-dom": "^6.6.1",
+    "react-share": "^4.4.1",
     "react-slick": "^0.29.0",
-    "react-test-renderer": "^17.0.2",
-    "react-timekeeper": "^2.2.1",
+    "react-test-renderer": "^18.2.0",
     "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.1",
+    "redux-thunk": "^2.4.2",
     "resize-observer-polyfill": "^1.5.1",
     "sass": "^1.35.2",
     "sass-loader": "^13.0.2",
@@ -130,45 +123,41 @@
     "stylelint-config-recommended-scss": "^5.0.2",
     "superagent": "^8.0.0",
     "terser-webpack-plugin": "^5.3.3",
-    "ts-loader": "^9.3.1",
-    "typescript": "^4.8.2",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
     "url-loader": "^4.1.1",
     "url-search-params-polyfill": "^8.1.1",
-    "use-persisted-state": "^0.3.3",
     "util": "^0.12.3",
-    "webpack": "^5.7.0",
-    "webpack-bundle-analyzer": "^4.6.1",
-    "webpack-cli": "^4.10.0"
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^5.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
-    "@types/redux-mock-store": "latest",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
     "babel-loader": "^8.2.3",
-    "enzyme": "latest",
-    "eslint": "^8.23.0",
+    "eslint": "^8.31.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-plugin-es": "^4.1.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.0.1",
+    "eslint-plugin-jest": "^27.2.0",
     "eslint-plugin-json": "^3.1.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.31.0",
+    "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-security": "^1.5.0",
-    "identity-obj-proxy": "latest",
-    "jest": "^29.0.1",
-    "redux-mock-store": "latest",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.3.1",
     "stylelint": "^14.2.0",
     "stylelint-config-standard": "^24.0.0",
     "stylelint-scss": "^4.1.0",
-    "ts-jest": "29.0.0-next.0",
-    "webpack-dev-server": "^4.8.1"
+    "ts-jest": "29.0.3",
+    "webpack-dev-server": "^4.11.1"
   },
   "snyk": true
 }
