@@ -1,7 +1,7 @@
 import { CodeResponse, googleLogout } from '@react-oauth/google';
 import { defaultAuth, Iauth } from 'src/providers/Auth.provider';
 import 'react-notifications-component/dist/theme.css';
-import jwt from 'jsonwebtoken';
+import jwt from 'jwt-simple';
 import superagent from 'superagent';
 import commonUtils from 'src/lib/commonUtils';
 
@@ -13,7 +13,7 @@ export interface GoogleBody {
 }
 
 const setUserAuth = async (
-  setAuth: (arg0:Iauth)=>void,
+  setAuth: (arg0: Iauth) => void,
   token: string,
   userId: string | undefined,
 ) => {
@@ -24,14 +24,14 @@ const setUserAuth = async (
   });
 };
 
-const setUser = async (auth:Iauth, setAuth:(args0:Iauth)=>void, token:string): Promise<void> => {
-  const { sub } = jwt.verify(token, process.env.HashString as string) as jwt.JwtPayload;
+const setUser = async (auth: Iauth, setAuth: (args0: Iauth) => void, token: string): Promise<void> => {
+  const { sub } = jwt.decode(token, process.env.HashString as string);
   await setUserAuth(setAuth, token, sub);
 };
 
 const authenticate = async (
   googleBody: GoogleBody,
-): Promise<{ token:string, email:string }> => {
+): Promise<{ token: string, email: string }> => {
   const { body } = await superagent.post(`${process.env.BackendUrl}/user/auth/google`)
     .set({ Accept: 'application/json' }).send(googleBody);
   return body;
@@ -44,8 +44,8 @@ const makeState = () => () => {
 
 const responseGoogleLogin = async (
   response: Omit<CodeResponse, 'error' | 'error_description' | 'error_uri'>,
-  auth:Iauth,
-  setAuth: (arg0:Iauth)=>void,
+  auth: Iauth,
+  setAuth: (arg0: Iauth) => void,
 ): Promise<void> => {
   try {
     const uri = window.location.href;
@@ -53,7 +53,7 @@ const responseGoogleLogin = async (
     const body = {
       clientId: process.env.GoogleClientId,
       redirectUri: !baseUri.includes('localhost')
-    && process.env.NODE_ENV === 'production' ? `https://${baseUri}`
+        && process.env.NODE_ENV === 'production' ? `https://${baseUri}`
         : `http://${baseUri}`,
       code: `${response.code}`,
       state: makeState(),
@@ -61,13 +61,14 @@ const responseGoogleLogin = async (
     const { token } = await authenticate(body);
     await setUser(auth, setAuth, token);
   } catch (e) {
+    console.log(e);
     const eMessage = (e as Error).message;
     commonUtils.notify('Failed to authenticate', eMessage, 'danger');
     setAuth({ ...defaultAuth, error: eMessage });
   }
 };
 
-const responseGoogleLogout = async (setAuth:(arg0:Iauth)=>void): Promise<void> => {
+const responseGoogleLogout = async (setAuth: (arg0: Iauth) => void): Promise<void> => {
   googleLogout();
   setAuth(defaultAuth);
   window.location.reload();
