@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { BrowserRouter, RouteComponentProps } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import type { ImenuItem } from 'src/App/AppTemplate/menuConfig';
 import { checkIsAllowed, continueMenuItem, SideMenuItem } from 'src/App/AppTemplate/SideMenuItem';
@@ -9,24 +9,23 @@ describe('SideMenuItem', () => {
   it('is defined', () => {
     expect(SideMenuItem).toBeDefined();
   });
-  it('returns MakeLink for /music', () => {
+  it('renders MakeLink for /music', () => {
     commonUtils.getUserRoles = jest.fn(() => ['tester']);
     const props = {
       index: 1,
       auth: { isAuthenticated: false, user: { userType: 'joker' } } as any,
-      location: { pathname: '/music' } as RouteComponentProps['location'],
       menu: { auth: false, link: '/music' } as ImenuItem,
       handleClose: jest.fn(),
     };
-    const smi: any = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>).toJSON();
-    expect(smi.props.className).toBe('menu-item');
+    const smi = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>).root;
+    expect(smi).toBeDefined();
   });
   it('returns MakeLink for /music when Web Jam LLC', () => {
     commonUtils.getUserRoles = jest.fn(() => ['tester']);
     const props = {
       index: 1,
       auth: { isAuthenticated: false, user: { userType: 'joker' } } as any,
-      location: { pathname: '/music' } as RouteComponentProps['location'],
+      location: { pathname: '/music' },
       menu: { auth: false, link: '/', name: 'Web Jam LLC' } as ImenuItem,
       handleClose: jest.fn(),
     };
@@ -46,8 +45,8 @@ describe('SideMenuItem', () => {
     expect(result.key).toBe('googleLogout');
   });
   it('checkIsAllowed return false if item requires auth and userType is not allowed', () => {
-    const menu:any = { auth: true };
-    const auth:any = { isAuthenticated: true, user: { userType: 'tester' } };
+    const menu: any = { auth: true };
+    const auth: any = { isAuthenticated: true, user: { userType: 'tester' } };
     const userRoles = ['admin'];
     const isAllowed = checkIsAllowed(menu, auth, userRoles);
     expect(isAllowed).toBe(false);
