@@ -8,23 +8,30 @@ import { BrowserRouter } from 'react-router-dom';
 
 describe('/music', () => {
   it('renders the component', () => {
-    const music = renderer.create(<Music />).toJSON();
+    const music = renderer.create(<BrowserRouter><Music /></BrowserRouter>).toJSON();
     expect(music).toMatchSnapshot();
   });
   it('renders and runs useEffect', () => {
     const container = document.createElement('div');
     document.body.appendChild(container);
     act(() => {
-      ReactDom.render(<Music />, container);
+      ReactDom.render(<BrowserRouter><Music /></BrowserRouter>, container);
     });
-    const musicSlider:any = document.getElementById('musicSlide1');
+    const musicSlider: any = document.getElementById('musicSlide1');
     expect(musicSlider).toBeDefined();
     document.body.removeChild(container);
   });
   it('renders ClickToListen and handles click when joshandmariamusic.com', () => {
     window.open = jest.fn();
     const ctl = renderer.create(
-      <ClickToListen appName="joshandmariamusic.com" isAdmin={false} setShowCreatePic={jest.fn()} setShowEditPic={jest.fn()} />,
+      <BrowserRouter>
+        <ClickToListen
+          appName="joshandmariamusic.com"
+          isAdmin={false}
+          setShowCreatePic={jest.fn()}
+          setShowEditPic={jest.fn()}
+        />
+      </BrowserRouter>,
     ).root;
     ctl.findByProps({ variant: 'contained' }).props.onClick();
     // eslint-disable-next-line security/detect-non-literal-fs-filename
@@ -57,7 +64,7 @@ describe('/music', () => {
     const setIsAdmin = jest.fn();
     const userRoles = process.env.userRoles || '{}';
     const { roles } = JSON.parse(userRoles);
-    const auth:any = { isAuthenticated: true, user: { userType: roles[0] } };
+    const auth: any = { isAuthenticated: true, user: { userType: roles[0] } };
     checkIsAdmin(auth, setIsAdmin);
     expect(setIsAdmin).toHaveBeenCalledWith(true);
   });
