@@ -1,5 +1,3 @@
-import { configure } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import { config } from 'dotenv';
 
 global.ResizeObserver = require('resize-observer-polyfill');
@@ -13,7 +11,6 @@ window.matchMedia = jest.fn().mockImplementation((query) => ({
   removeListener: jest.fn(),
 }));
 
-configure({ adapter: new Adapter() });
 document.body.innerHTML = '<div id="root"><div id="mAndP"></div><div id="play-buttons">'
   + '</div><div id="share-buttons"></div><div id="googleMap"></div></div>';
 window.HTMLMediaElement.prototype.load = () => { /* do nothing */ };
@@ -23,6 +20,7 @@ Object.defineProperty(window, 'location', { value: { assign: () => { }, reload: 
 window.location = {
   ...window.location,
   href: 'https://web-jam.com',
+  origin: 'https://web-jam.com',
   reload: jest.fn(),
   assign: jest.fn(),
 };
